+++ package.json
@@ -33,11 +33,6 @@
     "@openstreetmap/id-tagging-schema": "^6.12.0",
     "@sentry/nextjs": "^8.34.0",
     "@teritorio/openmaptiles-gl-language": "^1.5.4",
-    "@turf/buffer": "^7.2.0",
-    "@turf/concave": "^7.2.0",
-    "@turf/convex": "^7.2.0",
-    "@turf/helpers": "^7.2.0",
-    "@turf/polygon-smooth": "^7.2.0",
     "@turf/turf": "^7.2.0",
     "@xmldom/xmldom": "^0.9.8",
     "accept-language-parser": "^1.5.0",
