+++ src/components/Map/climbingTiles/climbingTilesSource.ts
@@ -15,13 +15,7 @@ import {
 import { join } from '../../../utils';
 import { mapClimbingFilter } from '../../utils/userSettings/getClimbingFilter';
 import { decodeHistogram } from '../../../server/climbing-tiles/overpass/histogram';
-import { Feature as GeojsonFeature, Polygon } from 'geojson';
-import { featureCollection, point } from '@turf/helpers';
-import convex from '@turf/convex';
-import polygonSmooth from '@turf/polygon-smooth';
-import buffer from '@turf/buffer';
-import { bbox } from '@turf/turf';
-import distance from '@turf/distance';
+import { constructOutlines } from './constructOutlines';
 
 const getTileJson = async ({ z, x, y }: Tile) => {
   try {
@@ -97,65 +91,6 @@ const doClimbingFilter = (features: ClimbingTilesFeature[]) => {
     return true;
   });
   return filteredFeatures;
-};
-
-const constructOutlines = (filteredFeatures: ClimbingTilesFeature[]) => {
-  return filteredFeatures
-    .filter(({ id }) => (id as number) % 10 === 4)
-    .flatMap((relation) => {
-      const mapId = relation.id as number;
-      const relationId = Math.floor(mapId / 10);
-
-      const subfeatures = filteredFeatures.filter(
-        (f) => f.properties.parentId === relationId,
-      );
-
-      // construct box around subfeatures
-      const coordinates = subfeatures.flatMap((f) => {
-        if (f.geometry.type === 'Point') {
-          return [f.geometry.coordinates];
-        }
-        if (f.geometry.type === 'LineString') {
-          return f.geometry.coordinates;
-        }
-        return [];
-      });
-
-      if (coordinates.length === 0) {
-        return [];
-      }
-
-      const fc = featureCollection(coordinates.map((c) => point(c)));
-      const hull = convex(fc);
-
-      if (!hull) {
-        return [];
-      }
-
-      // Compute bounding box of hull
-      const [minX, minY, maxX, maxY] = bbox(hull);
-      const width = maxX - minX;
-      const height = maxY - minY;
-      const maxDimension = Math.max(width, height);
-      const meters = distance([minX, minY], [maxX, maxY], { units: 'meters' });
-      const minZoom = Math.log2((5 * 40075016) / (meters * 256));
-
-      // Buffer by 10% of largest dimension in degrees (~approximation)
-      const bufferDistance = maxDimension * 0.1;
-      const buffered = buffer(hull, bufferDistance, { units: 'degrees' });
-      const smooth = polygonSmooth(buffered, { iterations: 3 });
-
-      return [
-        {
-          ...(smooth as any).features[0],
-          id: mapId,
-          properties: {
-            type: 'outline',
-            minZoom,
-          },
-        } as GeojsonFeature<Polygon>,
-      ];
-    });
 };
 
 const updateData = async () => {
